edited Security Configuration class with right expressions/maven

+++ src/main/java/leiden/iprwc/ecommerce/dao/ProductCategoryRepository.java
@@ -7,5 +7,5 @@ import org.springframework.web.bind.annotation.CrossOrigin;
 
 @RepositoryRestResource(collectionResourceRel = "productCategory", path = "product-category", excerptProjection = ProductCategory.class)
 public interface ProductCategoryRepository extends JpaRepository<ProductCategory, Long> {
-    ProductCategory findById(int id);
+    ProductCategory findById(long id);
 }
